Add files via upload
Please to use this library, here are some things to note; if you want to only use the display functionalities, use the library called display.py via the line "from display import *" but, if you want to use both the display and the touchscreen at the same time, use "from lowres.display import *". To use the touchscreen functionalities alone, please use the line "from touchscreen import *".

Use the folders and files that are the most necessary for your projects.